+++ editor_src/TextureFactory.h
@@ -12,7 +12,7 @@
 class TextureFactory{
     public:
         void unmarshalMap(const char *yamlFile, TextureMap& map, std::vector<std::unique_ptr<SdlTexture>>& textures,
-         std::vector<float>& mapSize, SdlRenderer& renderer);
+         std::vector<int>& mapSize, SdlRenderer& renderer);
         void unmarshalBombSites(const char *yamlFile, std::map<std::string, std::unique_ptr<Draggable>>& bombSites,
          SdlRenderer& renderer, const int& tile_size);
         void unmarshalSpawnSites(const char *yamlFile, std::map<std::string, std::unique_ptr<Draggable>>& bombSites,
